Move the queue job state enum into the message-queue module
JobState and its BullMQ mapping describe a queue concept but lived in the
admin-panel feature module, so reading job status from the application module
coupled two sibling feature modules. They now sit in message-queue, the lower
layer both already depend on. The GraphQL enum keeps its name and members, so
the generated schemas are unchanged.
